Added function to check for file before removal

+++ Login-script.sh
@@ -51,14 +51,20 @@ DEB_PACKAGE_NAME="python2.7 python2-dev libssl-dev"
 # Set Version 
 version="v0.0.1"
 export version="$version"
+if [ ! -f ~/Documents/.version ]; then
 sudo rm ~/Documents/.version
+else
 echo "$version" > ~/Documents/.version
+fi
 #-------------------------------------------------------------------------------------------------------------------------
 # Set the overall script name
 scriptname="RPH CnC BASH Menu"
 export scriptname="$scriptname"
+if [ ! -f ~/Documents/.scriptname ]; then
 sudo rm ~/Documents/.scriptname
+else
 echo "$scriptname" > ~/Documents/.scriptname
+fi
 #-------------------------------------------------------------------------------------------------------------------------
 # Get the filename for display
 me="$(basename "$(test -L "$0" && readlink "$0" || echo "$0")")"
